refactor(examples): move plan template to templates file
